Variant Rank Score Normalization
** Variant Ranking **

Normalize variant rank score by means of max-min normalization.

This is preferable because it makes output values
agnostic to the number of plugins and their scoring
behavior.

Max and min is computed as the total sum of upper and lower
bounds of each plugin.

The VCF attribute 'RankScoreNormalized' (bounded (0, 1))
is added to the scored VCF file.

No changes to existing 'RankScore' behavior, which
is still present in VCF.

** Compound Ranking **

This patch makes CompoundScore class use the additional
RankScoreNormalized VCF field for scoring purposes.
This patch adds a CompoundsNormalized INFO field to VCF,
based on RankScoreNormalized.

* Export RankScore bounds MIN, MAX to VCF file
  This is required by CompoundScore because it thresholds and
  edits RankScore(Normalized) during the scoring process.

* Configure CompoundScorer to use normalized rank score
  in addition to legacy RankScore.

* Apply CompoundScorer custom thresholds normalized
  to RankScore Min-Max bounds.

Signed-off-by: Tor Björgen <tor.bjorgen@scilifelab.se>